requirements.txt
requests
langchain
langchain-openai
streamlit
crewai
python-dotenv
gspread
oauth2client
waitress
